requirements.txt
langchain
openai
PyPDF2
cassio
langchain-community
python-dotenv
ipykernel 
pypdf
streamlit
sentence-transformers
langchain-huggingface
langchain-openai